create NonParallelizableUserJourneyPageTest, make SubmitPageTest an integration test
